separate swagger
