+++ src/env.ts
@@ -1,0 +1,1 @@
+export const PORT = process.env.PORT || 3000;
